+++ src/shared/sol-util.c
@@ -179,82 +179,6 @@ sol_util_strtodn(const char *nptr, char **endptr, ssize_t len, bool use_locale)
     return value;
 }
 
-#ifdef SOL_PLATFORM_CONTIKI
-#include <contiki.h>
-
-struct timespec
-sol_util_timespec_get_current(void)
-{
-    struct timespec ret;
-    clock_time_t ticks;
-
-    ticks = clock_time();
-    ret.tv_sec = ticks / CLOCK_SECOND;
-    ticks -= ret.tv_sec * CLOCK_SECOND;
-    ret.tv_nsec = (ticks * NSEC_PER_SEC) / CLOCK_SECOND;
-    return ret;
-}
-#elif defined(SOL_PLATFORM_RIOT) && SOL_PLATFORM_RIOT
-#include <vtimer.h>
-
-struct timespec
-sol_util_timespec_get_current(void)
-{
-    struct timespec tp;
-    timex_t t;
-    vtimer_now(&t);
-    tp.tv_sec = t.seconds;
-    tp.tv_nsec = t.microseconds * 1000;
-    return tp;
-}
-#else
-struct timespec
-sol_util_timespec_get_current(void)
-{
-    struct timespec t;
-
-    clock_gettime(CLOCK_MONOTONIC, &t);
-    return t;
-}
-#endif
-
-#ifdef SOL_PLATFORM_CONTIKI
-int
-sol_util_timespec_get_realtime(struct timespec *t)
-{
-    errno = ENOSYS;
-    return -1;
-}
-#elif defined(SOL_PLATFORM_RIOT) && SOL_PLATFORM_RIOT
-#if FEATURE_PERIPH_RTC
-#include <periph/rtc.h>
-#endif
-
-int
-sol_util_timespec_get_realtime(struct timespec *t)
-{
-#if FEATURE_PERITH_RTC
-    struct tm rtc;
-    if (rtc_get_time(&rtc) != 0) {
-        errno = EINVAL;
-        return -1;
-    }
-    t.tv_sec = mktime(&rtc);
-    t.tv_nsec = 0;
-    return 0;
-#else
-    errno = ENOSYS;
-    return -1;
-#endif
-}
-#else
-int
-sol_util_timespec_get_realtime(struct timespec *t)
-{
-    return clock_gettime(CLOCK_REALTIME, t);
-}
-#endif
-
 char *
 sol_util_strerror(int errnum, char *buf, size_t buflen)
 {
